commiting after a long time

+++ app/Http/Controllers/LinksController.php
@@ -4,22 +4,40 @@ namespace App\Http\Controllers;
 
 use App\Http\Controllers\Controller;
 use Illuminate\Support\Facades\DB;
+use Illuminate\Support\Facades\Validator;
 
 use Illuminate\Http\Request;
 use App\Models\Page;
 use App\Models\User;
 use App\Models\Link;
+use App\Models\Click;
 
 
 class LinksController extends Controller
 {
-    public function store(Request $request){
+    public function store(Request $request)
+    {
+        $validator = Validator::make($request->all(), [
+            'title' => 'required|string',
+            'description' => 'string',
+            'type' => 'required|integer',
+            'url' => 'required|regex:/^[-a-zA-Z0-9@:%._\\+~#=]{1,256}\\.[a-zA-Z0-9()]{1,6}\\b(?:[-a-zA-Z0-9()@:%_\\+.~#?&\\/=]*)$/',
+            'image' => $request->input('type') == 2
+             | $request->input('type') == 3 
+             | $request->input('type') == 4
+             ? 'required|mimes:jpeg,jpg,png|max:2048'
+             : 'nullable|mimes:jpeg,jpg,png|max:2048',
+        ]);
 
-        if($request->input('type') == 1){
+        if ($validator->fails()) {
+            return response()->json($validator->messages(), 422);
+        }
+
+        if ($request->input('type') == 1) {
             $user_id = $request->user()->id;
             $page = Page::where('user_id', $user_id)->first();
             $previous_link = Link::where('page_id', $page->id)->latest()->first();
-            if ($previous_link){
+            if ($previous_link) {
                 $previous_link_order = $previous_link->link_order;
                 $link_order = $previous_link_order + 1;
                 Link::create([
@@ -30,7 +48,7 @@ class LinksController extends Controller
                     'page_id' => $page->id,
                     'link_order' => $link_order
                 ]);
-            }else{
+            } else {
                 Link::create([
                     'title' => $request->input('title'),
                     'description' => $request->input('description'),
@@ -46,14 +64,14 @@ class LinksController extends Controller
                 $file = $request->file('image');
                 $filename = $file->getClientOriginalName(); //or you can give a name
                 $path = '/images';
-                $file->move(public_path($path),$filename);
-                $filename = 'images/'. $filename;
+                $file->move(public_path($path), $filename);
+                $filename = 'images/' . $filename;
             }
 
             $user_id = $request->user()->id;
             $page = Page::where('user_id', $user_id)->first();
             $previous_link = Link::where('page_id', $page->id)->latest()->first();
-            if ($previous_link){
+            if ($previous_link) {
                 $previous_link_order = $previous_link->link_order;
                 $link_order = $previous_link_order + 1;
                 Link::create([
@@ -65,7 +83,7 @@ class LinksController extends Controller
                     'page_id' => $page->id,
                     'link_order' => $link_order
                 ]);
-            }else{
+            } else {
                 Link::create([
                     'thumbnail_path' => $filename,
                     'title' => $request->input('title'),
@@ -79,11 +97,21 @@ class LinksController extends Controller
 
             return response('success');
         }
-       
     }
 
     public function update(Request $request)
     {
+
+        $validator = Validator::make($request->all(), [
+            'id' => 'required|string',
+            'title' => 'required|string',
+            'hyperlink' => 'required|url',
+        ]);
+
+        if ($validator->fails()) {
+            return response()->json(['error' => $validator->errors()], 422);
+        }
+
         if ($request->has('id')) {
             $link = Link::find($request->input('id'));
             $link->title = $request->input('title');
@@ -95,23 +123,77 @@ class LinksController extends Controller
 
     public function destroy(Request $request)
     {
-            $link = Link::find($request->input('id'));
-            $link->delete();
-            return back();
+
+        $validator = Validator::make($request->all(), [
+            'id' => 'required|integer',
+        ]);
+
+        if ($validator->fails()) {
+            return response()->json(['error' => $validator->errors()], 422);
+        }
+
+        $link = Link::find($request->input('id'));
+        $link->delete();
+        return back();
     }
 
-    public function getLinks(Request $request) {  
+    public function getLinks(Request $request)
+    {
         $user_id = $request->user()->id;
         $page = Page::where('user_id', $user_id)->first();
         $links = Link::where('page_id', $page->id)->orderBy('link_order', 'asc')->get()->toArray();
         return ['links' => $links];
     }
 
-    public function move(Request $request, Link $link){
+    public function move(Request $request, Link $link)
+    {
+
+        $validator = Validator::make($request->all(), [
+            'link_order' => 'required|integer',
+        ]);
+
+
         $link->update([
             'link_order' => round($request->input('link_order'), 5)
         ]);
     }
 
+    public function linkClick(Request $request)
+    {
+        $validator = Validator::make($request->all(), [
+            'id' => 'required|integer',
+        ]);
 
+        if ($validator->fails()) {
+            return response()->json(['error' => $validator->errors()], 422);
+        }
+
+        $link = Link::where('id', $request->input('id'))->first();
+        $link->clicks()
+            ->create([
+                'link_id' => $request->input('id'),
+                'user_agent' => $request->userAgent(),
+                'ip' => $request->ip()
+            ]);
+    }
+
+    public function getTotalClicks(Request $request)
+    {
+        $user_id = $request->user()->id;
+        $page = Page::where('user_id', $user_id)->first();
+        $links = Link::where('page_id', $page->id)->get();
+        $count = 0;
+        foreach ($links as $link) {
+            $count += self::singleLinkClickCount($link->id);
+        }
+        return $count;
+    }
+
+    public function singleLinkClickCount($id)
+    {
+        $link = Link::where('id', $id)->first();
+        $click = Click::where('link_id', $link->id)->get();
+        $count = $click->count();
+        return $count;
+    }
 }
